update Quote
Id veranderd van int naar Long

+++ src/main/java/nl/mengelmoestuintjes/gardening/repository/QuoteRepository.java
@@ -3,5 +3,5 @@ package nl.mengelmoestuintjes.gardening.repository;
 import nl.mengelmoestuintjes.gardening.model.Quote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuoteRepository extends JpaRepository<Quote, Integer> {
+public interface QuoteRepository extends JpaRepository<Quote, Long> {
 }
